+++ pkg/metrics/metrics.go
@@ -1,17 +1,24 @@
 package metrics
 
-import "github.com/prometheus/client_golang/prometheus"
+import (
+	"fmt"
 
-const namespace = "neofs_node"
+	"github.com/prometheus/client_golang/prometheus"
+)
+
+const storageNodeNameSpace = "neofs_node"
 
 type NodeMetrics struct {
 	objectServiceMetrics
 	engineMetrics
 	stateMetrics
-	epoch prometheus.Gauge
+	epoch           prometheus.Gauge
+	epochDeprecated prometheus.Gauge
 }
 
-func NewNodeMetrics() *NodeMetrics {
+func NewNodeMetrics(version string) *NodeMetrics {
+	registerVersionMetric(storageNodeNameSpace, version)
+
 	objectService := newObjectServiceMetrics()
 	objectService.register()
 
@@ -22,12 +29,22 @@ func NewNodeMetrics() *NodeMetrics {
 	state.register()
 
 	epoch := prometheus.NewGauge(prometheus.GaugeOpts{
-		Namespace: namespace,
-		Subsystem: innerRingSubsystem,
+		Namespace: storageNodeNameSpace,
+		Subsystem: stateSubsystem,
 		Name:      "epoch",
-		Help:      "Current epoch as seen by inner-ring node.",
+		Help:      "Current epoch as seen by NeoFS node.",
 	})
 	prometheus.MustRegister(epoch)
+
+	// FIXME: drop after v0.38.0 release: #2347.
+	const stateSubsystemDeprecated = objectSubsystem
+	epochDeprecated := prometheus.NewGauge(prometheus.GaugeOpts{
+		Namespace: storageNodeNameSpace,
+		Subsystem: stateSubsystemDeprecated,
+		Name:      "epoch",
+		Help:      fmt.Sprintf("Current epoch as seen by inner-ring node. DEPRECATED: use [%s_%s_epoch] instead.", storageNodeNameSpace, stateSubsystem),
+	})
+	prometheus.MustRegister(epochDeprecated)
 
 	return &NodeMetrics{
 		objectServiceMetrics: objectService,
@@ -40,4 +57,5 @@ func NewNodeMetrics() *NodeMetrics {
 // SetEpoch updates epoch metric.
 func (m *NodeMetrics) SetEpoch(epoch uint64) {
 	m.epoch.Set(float64(epoch))
+	m.epochDeprecated.Set(float64(epoch))
 }